update: ignore console test in githug action

+++ scripts/e2e_test/start_test.sh
@@ -249,7 +249,9 @@ api_test() {
 }
 
 console_test() {
-  docker run --rm --ipc=host -w /app -e PROXY=${SWNAME//./-}.pre.intra.starwhale.ai -v $SWROOT/console/playwright:/app mcr.microsoft.com/playwright:v1.33.0-jammy /bin/bash -c "yarn && yarn test" || exit 1
+  if ! in_github_action; then
+    docker run --rm --ipc=host -w /app -e PROXY=${SWNAME//./-}.pre.intra.starwhale.ai -v $SWROOT/console/playwright:/app mcr.microsoft.com/playwright:v1.33.0-jammy /bin/bash -c "yarn && yarn test" || exit 1
+  fi
 }
 
 restore_env() {
